Changes the batching function so that we process <batch_size> pairs of inputs per batch, rather than <nb_queriers_per_batch> X <per_query_nb_examples>. Also modifies the scoring function of the model so that we must feed encodings, and these can be either 1D or 2D tensors.

+++ tr4hd/data_utils.py
@@ -271,7 +271,7 @@ def make_q_and_c_dataset(opt, tokenizer, queries, candidate_ids, candidate_label
             logger.info("  candidate ids: %s" % " ".join([str(x) for x in candidate_ids[i]]))
             logger.info("  candidate labels: %s" % " ".join([str(x) for x in candidate_labels[i]]))
     candidate_ids = torch.tensor(candidate_ids, dtype=torch.long)
-    candidate_labels = torch.tensor(candidate_labels, dtype=torch.long)
+    candidate_labels = torch.tensor(candidate_labels, dtype=torch.float32)
     return TensorDataset(input_ids, nb_tokens, candidate_ids, candidate_labels)
